Unify slide rendering on one merge-based path; drop per-type leaf views
Having built both, the per-type useFragment leaf views (TextView/ImageView/…)
were a second render path that re-implemented cmpStyle/renderInner and bought
nothing here: the fragments are shared with the editor so every column is
selected (no wire trimming), and the editor is intrinsically cross-type
(select / marquee / z-order / inspector all walk one merged list).

Keep the real win — the composed deckDetailQuery + SlideFragment (one query
per deck, SSR-able) — and unify everything on a single render path:
useFragment(SlideFragment) -> mergeComponents -> shared cmpStyle/renderInner.
The editor's Stage wraps each component with selection/resize handles; the
read-only SlideView renders the static box (black placeholders for video/web
frames). Same data path, same renderer, no duplication.

- ComponentViews.tsx (5 leaf views + SlideView) -> SlideView.tsx (merge-based)
- Stacking back to DOM order (mergeComponents sorts by z_order), matching the
  proven pre-relay thumbnail behavior
- Update the 4 import sites + Stage's data-path comment

tsc clean; client+SSR build green; eslint 57 -> 42 (new file fully clean).
