+++ src/routes/deck.$deckId_.play.tsx
@@ -2,7 +2,7 @@ import { createFileRoute, useNavigate } from '@tanstack/react-router'
 import { useEffect, useLayoutEffect, useRef, useState } from 'react'
 import { useQuery } from '@rindle/react'
 import { deckDetailQuery } from '../../shared/queries'
-import { SlideView } from '../editor/ComponentViews'
+import { SlideView } from '../editor/SlideView'
 import { resolveSurface } from '../editor/types'
 import { UserStyle } from '../editor/CssEditor'
 import { flightFor } from '../editor/transitions'
