+++ src/routes/share.$deckId.tsx
@@ -7,7 +7,7 @@ import { createFileRoute } from '@tanstack/react-router'
 import { useEffect, useLayoutEffect, useState } from 'react'
 import { useQuery } from '@rindle/react'
 import { publicDeckDetailQuery } from '../../shared/queries'
-import { SlideView } from '../editor/ComponentViews'
+import { SlideView } from '../editor/SlideView'
 import { resolveSurface } from '../editor/types'
 import { UserStyle } from '../editor/CssEditor'
 import { flightFor } from '../editor/transitions'
